requirements.txt
astroid==2.12.13
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
dill==0.3.6
dnspython==2.2.1
idna==3.4
isort==5.11.3
lazy-object-proxy==1.8.0
loguru==0.6.0
mccabe==0.7.0
oauthlib==3.2.2
platformdirs==2.6.0
pycodestyle==2.10.0
pylint==2.15.9
pymongo==4.3.3
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
tomli==2.0.1
tomlkit==0.11.6
tweepy==4.12.1
urllib3==1.26.13
wrapt==1.14.1
